Add files via upload
For this project I have been asked to write a program for a retail store that will allow them to calculate discounts for their employees when they buy items. Discounts are based on number of years worked as well as if the employee is a manager or hourly employee. They are also allowed no discount once they have received $200 in discounts for the year. Once all employees have been processed, I display the "All Employee Summary" and end the program.